% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLambda.R
\name{applyLambda}
\alias{applyLambda}
\title{applyLambda}
\usage{
applyLambda(matrix_wsm, matrix_wpm, lambda)
}
\arguments{
\item{matrix_wsm}{The data set object obtained from the application of the}

\item{matrix_wpm}{The data set object obtained from the application of the}

\item{lambda}{The lambda value (between 0 and 1)}
}
\value{
A data frame object that contains the alternatives set scored by and
  classified in descending
        order (from best to worst classified) according to the weighting
        proposed by the WASPAS method
        using the input lambda.
}
\description{
Apply the lambda to assign a relative importance to each of the
  previously used methods (WSM and WPM). Lambda values range from zero to one
}
\examples{

\donttest{
applyLambda(matrix_wsm, matrix_wpm, lambda)
waspas_rank <- applyLambda(matrix_wsm, matrix_wpm, lambda)
waspas_rank <- applyLambda(matrix_wsm, matrix_wpm, 0)
waspas_rank <- applyLambda(matrix_wsm, matrix_wpm, 0.5)
waspas_rank <- applyLambda(matrix_wsm, matrix_wpm, 0.99)
}
}
