% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWSM.R
\name{calcWSM}
\alias{calcWSM}
\title{calcWSM}
\usage{
calcWSM(AxCNorm, vWeights)
}
\arguments{
\item{AxCNorm}{A data set object with normalized values of Alternatives X Criteria}

\item{vWeights}{Contains a set of user-assigned values to weight the criteria.
The sum of these weights must add up to 1.
The format of this input is an array of values.}
}
\value{
A data frame object that contains 2 columns and the the same number of rows as
       the input matrix. The columns "Points" has the calculated relative value of each
       alternative whose id is in the "Alternatives" column.
}
\description{
Calculates the ranking for the alternative's set according to WSM method.
}
\examples{
Do not just copy and paste!

\dontrun{
calcWSM(normalized_matrix, vector_weights)
wpm_matrix <- calcWSM(normalized_matrix, vector_weights)}
}
