% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waspasR.R
\name{waspasR}
\alias{waspasR}
\title{waspasR()}
\usage{
waspasR(waspas_df, lambda)
}
\arguments{
\item{waspas_df}{The original data set in a proper format. The format can be
checked by checkInputFormat() function.}

\item{lambda}{The lambda value (between 0 and 1)}
}
\value{
A data.frame object that contains the input matrix with its values
   normalized. Or an error message if some bad data is entered.
}
\description{
Runs the complete process from slicing the original database,
   processing all the computational steps, like computing WSM and WPM
   formulas, applying the lambda as proposed by the method WASPAS, and the
   building the complete output in a new data.frame with the criteria as
   column names, all the original data and appending 3 new columns with
   the WSM, WPM and WASPAS ranking ("WSM_Rank", "WPM_Rank", "WASPAS_Rank").
}
\examples{

\dontrun{
waspas_set <- waspasR(chopper, lambda = 0.23)
waspas_set <- waspasR(myRawData, lambda = 0.8)
}
}
