% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{checkInputFormat}
\alias{checkInputFormat}
\title{checkInputFormat}
\usage{
checkInputFormat(waspas_db)
}
\arguments{
\item{waspas_db}{The original database to be validated in its format}
}
\value{
True if everything is OK, an error message in case of bad format
}
\description{
Verify if the database to be submitted to WASPAS is correctly
formatted
}
