% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputFormat.R
\name{checkInputFormat}
\alias{checkInputFormat}
\title{checkInputFormat}
\usage{
checkInputFormat(dfMatrix)
}
\arguments{
\item{dfMatrix}{The original database to be validated in its format}
}
\value{
True if everything is OK, an error message in case of bad format
}
\description{
Verify if the database to be submitted to WASPAS is correctly
formatted
}
\examples{

\dontrun{
checkDF <- checkInputFormat(dfMatrix)
checkInputFormat(dfMatrix)
paste("My data.frame format:", checkInputFormat(dfMatrix))
}
}
