% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliceData.R
\name{sliceData}
\alias{sliceData}
\title{sliceData}
\usage{
sliceData(dfMatrix, outData)
}
\arguments{
\item{dfMatrix}{A matrix or data.frame in “all-in-one” format}

\item{outData}{A flag to determine the vector or matrix (data.frame) to extract from the input matrix
must be 'A' (Alternatives), 'C' (Criteria), 'F' (Flags), 'V' (Values) or 'W' (Weights)}
}
\value{
A data.frame one-dimensional (vector) or two-dimensional (matrix) with one of the Following objects:
 - if outData == "A": A vector of Alternatives
 - if outData == "C": A vector of Criteria
 - if outData == "F": A vector of Cost-Benefit Flags
 - if outData == "V": A matrix of values per Alternative x Criterion
 - if outData == "W": A vector of Weights
}
\description{
Slice a matrix or data.frame in “all-in-one” format into dedicated vectors/matrices
             as data.frame objects
}
\examples{

\dontrun{
AlternativesList <- sliceData(dfMatrix, "A")
CriteriaList <- sliceData(dfMatrix, "C")
cost_benefit_Flags <- sliceData(dfMatrix, "F")
values_matrix <- sliceData(dfMatrix, "M")
vectorWeights <- sliceData(dfMatrix, "W")}

}
