% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waspasR.R
\name{waspasR}
\alias{waspasR}
\title{waspasR()}
\usage{
waspasR(dfMatrix, lambda)
}
\arguments{
\item{dfMatrix}{A data set object with Alternatives X Criteria  values to be normalized}

\item{vCostBenefit}{A vector of flags that determines if the criterion is a Cost or Benefit one
Must be same size of Criteria, must contains just strings initiated with B, b, C or c}

\item{dataFormat}{Defines how the functions will deal with the input values. There are two possible values:
"All-in-One" (wich is the same that the values 1) or "simple"  (wich is the same as 2),
if ommited the function will assume the All-in-one format.}
}
\value{
A data frame object that contains the input matrix with its values normalized.
}
\description{
Normalize the values of Alternatives X Criteria matrix according to a given
             Cost-Benefit vector of flags.
             There are two ways to input the values:
             (a) The "All in One" format is a data.frame matrix with all the WSM, WPM and WASPAS data
             and metadata, in this case the parameter vCostBenefit can be null.
             (b) The "Simple" format is a data.frame matrix with just the values of Alternatives x Criteria,
             in this case the vector of Cost-Benefit flags must be entered in the parameter vCostBenefit.
             and metadata.
}
